Added a new dynamic tree shaking model to upb, with the intention of removing the old model once YouTube has migrated.
The `kUpb_DecodeOption_ExperimentalAllowUnlinked` flag to the decoder will enable the new behavior.  When that flag is not passed, tree shaking with the old model will still be possible.

"Dynamic tree shaking" in upb is a feature that allows messages to be parsed even if the MiniTables have not been fully linked.  Unlinked sub-message fields can be parsed by preserving their data in the unknown fields.  If the application later discovers that the message field is actually needed, the MiniTable can be patched to properly link that field, and existing message instances can "promote" the data from the unknown fields to an actual message of the correct type.

Before this change, dynamic tree shaking stored unparsed message data in the unknown fields of the *parent*.  In effect, we were treating the field as if it did not exist at all.  This meant that parsing an unlinked field did not affect the hasbits or oneof cases of the parent, nor did it create a `upb_Array` or `upb_Map` for array/map fields.  Only when a message was linked and promoted did any of these things occur.

While this model had some amount of conceptual simplicity, it caused significant problems with oneofs.  When multiple fields inside a single oneof are parsed from the wire, order matters, because later oneof fields must overwrite earlier ones.  Dynamic tree shaking can mean that some fields in a oneof are linked while others are not.  It is essential that we preserve this ordering semantic even when dynamic tree shaking is being used, but it is difficult to do if the oneof's data can be split between linked fields (which have been reified into parsed field data) and unlinked fields (whose data lives in the unknown fields of the parent).

To solve this problem, this CL changes the representation for unlinked fields.  Instead of being placed in the parent's unknown fields, we create an actual message instance for each unlinked message we parse, but we use a placeholder "empty message" MiniTable as the message's type.  All of the message's data will therefore be placed into the "empty message's" unknown fields.  But unlike before, this "empty message" is actually present according to the hasbits, oneof case, and `upb_Array`/`upb_Map` of the parent.  This means that all of the oneof presence logic works as normal.

Since the MiniTable can be patched at any time, we need a bit in the message instance itself to signal whether a pointer to a sub-message is an "empty message" or not.  When dynamic tree shaking is in use, all users must be capable of recognizing an empty message and acting accordingly (promoting, etc) even if the MiniTable itself says that the field is linked.

Because dynamic tree shaking imposes this extra requirement on users, we require that users pass an extra option to the decoder to allow parsing of unlinked sub-messages.  Many existing users of upb (Ruby, PHP, Python, etc) will always have fully-linked MiniTables, so there is no reason for them to add extra logic to handle empty messages.  By omitting the `kUpb_DecodeOption_ExperimentalAllowUnlinked` option, they will be relieved of the duty to check the tagged pointer that would indicate an empty, unlinked message.

For existing users of dynamic tree shaking, there are three main changes:

1. The APIs in message/promote.h have changed, and users will need to update to the new interfaces.

2. The model for maps has changed slightly.  Before, we required that map entries always had their values linked; for dynamic tree shaking to apply to maps, we required that the *entry* was left unlinked, not the entry's value.  In the new model, that is reversed: map entries must always be linked, but a map entry's value can be unlinked.

3. The presence model for unlinked fields has changed.  Unlinked fields will now register as "present" from the perspective of hasbits, oneof cases, and array/map entries.  Users must test the tagged pointer to know if a message is of the correct, linked type or whether it is a placeholder "empty" message.  There is a new function `upb_Message_GetTaggedMessagePtr()`, as well as a new accessor `upb_MessageValue.tagged_msg_val` that can be used to read and test the tagged pointer directly.

PiperOrigin-RevId: 535288031