feat: workaround for `DOMAIN` macro (#12903)
This is a macro on some (older) versions of GCC, and macOS, and Windows. Sigh. I moved the `#undef` block to a common section. I also took the opportunity to add a regression test for all these macros that need to be `#undef`'d.

Part of the work for googleapis/google-cloud-cpp#8125

Closes #12903

PiperOrigin-RevId: 535649278